Separate new DeepReplace to be used only in reports
