+++ app/utils/common.ts
@@ -26,11 +26,8 @@ export function sortByOrder<T extends { order: number }>(data: T[] | undefined) 
     return [...data].sort((a, b) => compareNumber(a.order, b.order));
 }
 
-export function reorder<T>(
-    data: T[],
-    orderKey = 'order',
-) {
-    return data.map((v, i) => ({ ...v, [orderKey]: i + 1 }));
+export function reorder<T extends { order?: number }>(data: T[]) {
+    return data.map((v, i) => ({ ...v, order: i + 1 }));
 }
 
 export function breadcrumb(args: (string | undefined)[], symbol = ' » ') {
