+++ app/utils/types.ts
@@ -8,21 +8,14 @@ export type DeepMandatory<T, K extends keyof any> = T extends object ? (
     )
 ) : T
 
-type DeepNonNullable<T> = T extends object ? (
-    T extends (infer K)[] ? (
-        DeepNonNullable<K>[]
-    ) : (
-        { [P in keyof T]-?: DeepNonNullable<T[P]> }
-    )
-) : NonNullable<T>;
-
 export type DeepReplace<T, A, B> = (
-    DeepNonNullable<T> extends DeepNonNullable<A>
+    T extends A
         ? B
         : (
             T extends (infer Z)[]
                 ? DeepReplace<Z, A, B>[]
                 : (
+
                     T extends object
                         ? { [K in keyof T]: DeepReplace<T[K], A, B> }
                         : T
